bulk -n: Don't remove .building dir unless this ran made it

+++ src/share/poudriere/common.sh
@@ -1107,7 +1107,8 @@ exit_handler() {
 		if [ ${CREATED_JLOCK:-0} -eq 1 ]; then
 			update_stats >/dev/null 2>&1 || :
 		fi
-		if [ ${DRY_RUN} -eq 1 ] && [ -n "${PACKAGES_ROOT}" ]; then
+		if [ ${DRY_RUN} -eq 1 ] && [ -n "${PACKAGES_ROOT}" ] &&
+		    [ ${PACKAGES_MADE_BUILDING:-0} -eq 1 ] ; then
 			rm -rf "${PACKAGES_ROOT}/.building" || :
 		fi
 	fi
@@ -1867,6 +1868,7 @@ stash_packages() {
 		# set; Must stay on the same device for linking.
 
 		mkdir -p ${PACKAGES}/.building
+		PACKAGES_MADE_BUILDING=1
 		# hardlink copy all top-level directories
 		find ${PACKAGES}/.latest/ -mindepth 1 -maxdepth 1 -type d \
 		    ! -name .building | xargs -J % cp -al % ${PACKAGES}/.building
